Add authenticate_extract.[hc]pp for functions to extract schame of users Fix template operator() : moved definition of .hpp Add DB_IS_STRING and used it in authenticate.h Move some structures defitions and constants from authenticate.c to authenticate.h to use them authenticate_extract.cpp Change signature of extractobjects Moved some functions (schema extraction) from trigger_manager.[ch] to trigger_description.[ch]pp
