+++ src/executables/unloaddb.h
@@ -31,7 +31,7 @@
 
 
 
-
+struct extract_context;
 class extract_output;
 
 extern char *database_name;
@@ -64,7 +64,7 @@ extern int extract_triggers (extract_context &ctxt, extract_output &output_ctx);
 extern int extract_triggers_to_file (extract_context &ctxt, const char *output_filename);
 extern int extract_indexes_to_file (extract_context &ctxt, const char *output_filename);
 extern int extract_classes (extract_context &ctxt, extract_output &schema_output_ctx);
-extern int extractobjects (const char *exec_name);
+extern int extractobjects (const char *exec_name, const char *output_dirname, const char *output_prefix);
 
 extern int create_filename_schema (const char *output_dirname, const char *output_prefix,
                                    char *output_filename_p, const size_t filename_size);
