+++ src/object/trigger_description.hpp
@@ -29,8 +29,9 @@
 #endif //defined(SERVER_MODE)
 
 #include <stdio.h>
-
+#include "dbtype_def.h"
 struct db_object;
+class extract_output;
 
 /*
  * TRIGGER_HELP
@@ -61,5 +62,8 @@ struct trigger_description
 
   void fprint (FILE *file);
 };
+
+int tr_dump_trigger (extract_output &output_ctx, DB_OBJECT * trigger_object);
+int tr_dump_selective_triggers (extract_output &output_ctx, DB_OBJLIST * classes);
 
 #endif // _TRIGGER_DESCRIPTION_HPP_
