+++ src/object/trigger_manager.h
@@ -196,7 +196,6 @@ typedef enum
   TR_DECISION_DO_NOT_CONTINUE
 } TR_RECURSION_DECISION;
 
-class extract_output;
 /* TRIGGER OBJECT ATTRIBUTES */
 /*
  * Names of the trigger class and its attributes.
@@ -371,12 +370,9 @@ extern void tr_invalidate_user_cache (void);
 extern const char *tr_time_as_string (DB_TRIGGER_TIME time);
 extern const char *tr_event_as_string (DB_TRIGGER_EVENT event);
 extern const char *tr_status_as_string (DB_TRIGGER_STATUS status);
-extern int tr_dump_trigger (extract_output &output_ctx, DB_OBJECT * trigger_object);
 #if defined(ENABLE_UNUSED_FUNCTION)
 extern int tr_dump_all_triggers (FILE * fp, bool quoted_id_flag);
 #endif
-extern int tr_dump_selective_triggers (extract_output &output_ctx, DB_OBJLIST * classes);
-
 extern void tr_free_trigger_list (TR_TRIGLIST * list);
 extern const char *tr_get_class_name (void);
